+++ backend/src/global.d.ts
@@ -1,0 +1,10 @@
+export {};
+declare global {
+    namespace Express {
+        export interface User {
+            _id: string;
+            username: string;
+            email: string;
+        }
+    }
+}
